requirements.txt
aiohttp==3.8.5
bip-utils==2.6.0
python-dotenv==1.0.0
requests==2.31.0
